add central_status_changed_observer.c and revoke cmake magic
