+++ CMakeLists.txt
@@ -17,10 +17,6 @@ if(CONFIG_SHIELD_PROSPECTOR_ADAPTER)
         endif()
 
         zephyr_library_sources(src/events/split_central_status_changed.c)
-        set_source_files_properties(
-                ${APPLICATION_SOURCE_DIR}/src/split/bluetooth/central.c
-                TARGET_DIRECTORY app
-                PROPERTIES HEADER_FILE_ONLY ON)
-        zephyr_library_sources(src/split/bluetooth/central.c)
+        zephyr_library_sources(src/split/bluetooth/central_status_changed_observer.c)
 
 endif()
